tests stochastic approx
